requirements.txt
sentence-transformers
faiss-cpu
numpy
fastapi
uvicorn
pydantic
torch
torchvision
pillow
jinja2
python-multipart